remove old client code and replace with vuetify/vue frontend.
