+++ client/Dockerfile
@@ -5,6 +5,5 @@ RUN npm install
 COPY . .
 RUN npm run build
 
-FROM nginx:1.18.0-alpine as production-stage
+FROM alpine:latest
 COPY --from=build-stage /opt/client/dist /opt/client
-COPY nginx.conf /etc/nginx/nginx.conf
