+++ client/src/components/LinePlot.vue
