+++ client/src/components/Plots.vue
