+++ client/src/components/VisitLabelSelector.vue
