+++ client/src/main.ts
@@ -1,10 +1,14 @@
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-// import store from "./store";
+import Vue from 'vue'
+import App from './App.vue'
+// import router from './router'
+import { store } from './store'
+import vuetify from './plugins/vuetify'
+
+Vue.config.productionTip = false
 
 new Vue({
-  router,
-  // store,
+  // router,
+  store,
+  vuetify,
   render: h => h(App)
-}).$mount("#app");
+}).$mount('#app')
