+++ client/tsconfig.json
@@ -12,9 +12,13 @@
     "allowSyntheticDefaultImports": true,
     "sourceMap": true,
     "baseUrl": ".",
+    "typeRoots": [
+      "./node_modules/@types",
+      "./node_modules/vuetify/types"
+    ],
     "types": [
       "webpack-env",
-      "jest"
+      "vuetify"
     ],
     "paths": {
       "@/*": [
