--- seq2view-client/src/shims-vue.d.ts
@@ -1,4 +1,0 @@
-declare module '*.vue' {
-  import Vue from 'vue'
-  export default Vue
-}
